tg_bot/requirements.txt
pydantic
aiogram
python-dotenv